requirements.txt
numpy
Pillow
seam-carving
streamlit
